requirements.txt
# Web framework
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9

# Data validation
pydantic==2.7.1
pydantic-settings==2.3.0

# ML / NLP
torch==2.3.0
transformers==4.41.2
sentence-transformers==3.0.1
scikit-learn==1.5.0
numpy==1.26.4
pandas==2.2.2

# Vector search
faiss-cpu==1.8.0

# Database (lightweight, no setup needed)
aiosqlite==0.20.0
sqlalchemy==2.0.30

# HTTP client (to call Ollama)
httpx==0.27.0

# Environment variables
python-dotenv==1.0.1

# Utilities
tqdm==4.66.4